refactor: move AcceptToMemoryPool into validation
Move the ~245-line AcceptToMemoryPool from main.cpp into validation.cpp
next to the rest of the transaction/block validation logic, and its
declaration from main.h into validation.h (issue #3125, workstream C2).
This is the natural next phase of the mempool modernization tracked in
issue #3029. Pure code move, no behavior change.

validation.cpp gains the two includes the moved code needs that it did
not already have (gridcoin/contract/contract.h for ValidateContracts,
policy/policy.h for IsStandardTx) and validation.h a CTxMemPool forward
declaration. One unqualified map<> in the moved body is std::-qualified
(main.cpp has using namespace std, validation.cpp does not -- the same
class of fix as the A2/A3 extractions). Callers are unaffected -- main.h
includes validation.h, so the declaration still resolves transitively.

main.cpp -252 lines (now 430 lines, from ~3,520 pre-#3030).
