+++ src/main.h
@@ -116,12 +116,8 @@ bool IsInitialBlockDownload();
 // only the wallet derives from it (issue #3125, workstream C3).
 bool OutOfSyncByAge();
 
-/** (try to) add transaction to memory pool **/
-bool AcceptToMemoryPool(CTxMemPool& pool, CTransaction &tx,
-                        CValidationState& state, bool* pfMissingInputs,
-                        int64_t entry_time = 0, bool test_only = false,
-                        CAmount* fee_out = nullptr, size_t* vsize_out = nullptr)
-    EXCLUSIVE_LOCKS_REQUIRED(cs_main);
+// AcceptToMemoryPool lives in validation.h, included above (issue #3125,
+// workstream C2).
