+++ src/validation.cpp
@@ -10,6 +10,7 @@
 #include "main.h"
 #include "validationinterface.h"
 #include "gridcoin/beacon.h"
+#include "gridcoin/contract/contract.h"
 #include "gridcoin/contract/registry.h"
 #include "gridcoin/claim.h"
 #include "gridcoin/consensus/block_rewards.h"
@@ -27,6 +28,7 @@
 #include "node/chainman.h"
 #include "node/orphan_blocks.h"
 #include "policy/fees.h"
+#include "policy/policy.h"
 #include "serialize.h"
 #include "util.h"
 #include "util/time.h"
@@ -1809,4 +1811,251 @@ bool GetTransaction(const uint256 &hash, CTransaction &tx, uint256 &hashBlock)
         }
     }
     return false;
+}
+
+// Mempool acceptance moved here from main.cpp (issue #3125, workstream C2);
+// the natural next phase of the mempool modernization tracked in #3029.
+bool AcceptToMemoryPool(CTxMemPool& pool, CTransaction &tx, CValidationState& state, bool* pfMissingInputs, int64_t entry_time, bool test_only, CAmount* fee_out, size_t* vsize_out) EXCLUSIVE_LOCKS_REQUIRED(cs_main)
+{
+    AssertLockHeld(cs_main);
+    if (pfMissingInputs)
+        *pfMissingInputs = false;
+
+    // Mandatory switch to binary contracts (tx version 2):
+    if (tx.nVersion < 2) {
+        return state.DoS(100, error("AcceptToMemoryPool : legacy transaction"), "bad-txns-version");
+    }
+
+    if (!CheckTransaction(tx, state))
+        return error("AcceptToMemoryPool : CheckTransaction failed");
+
+    // Coinbase is only valid in a block, not as a loose transaction
+    if (tx.IsCoinBase())
+        return state.DoS(100, error("AcceptToMemoryPool : coinbase as individual tx"), "coinbase");
+
+    // ppcoin: coinstake is also only valid in a block, not as a loose transaction
+    if (tx.IsCoinStake())
+        return state.DoS(100, error("AcceptToMemoryPool : coinstake as individual tx"), "coinstake");
+
+    // Rather not work on nonstandard transactions
+    if (!IsStandardTx(tx))
+        return state.Invalid(error("AcceptToMemoryPool : nonstandard transaction type"), "tx-nonstandard");
+
+    // Perform contextual validation for any contracts:
+
+    int DoS = 0;
+    if (!tx.GetContracts().empty() && !GRC::ValidateContracts(tx, DoS)) {
+        return state.DoS(DoS, error("%s: invalid contract in tx %s, assigning DoS misbehavior of %i",
+                                 __func__,
+                                 tx.GetHash().ToString(),
+                                 DoS),
+                         "invalid-contract");
+    }
+
+    // is it already in the memory pool?
+    uint256 hash = tx.GetHash();
+    if (pool.exists(hash))
+        return false;
+
+    // Reject a second MRC for a CPID that already has one in the mempool. This
+    // is an O(log n) lookup against the m_mrc_by_cpid index (it replaced a full
+    // mempool scan + bloom filter). MRC DoS protection: a duplicate earns a
+    // stiff penalty.
+    bool tx_contains_valid_mrc = false;
+
+    for (const auto& contract : tx.GetContracts()) {
+        if (contract.m_type == GRC::ContractType::MRC) {
+            GRC::MRC mrc = contract.CopyPayloadAs<GRC::MRC>();
+            GRC::Cpid cpid = *(mrc.m_mining_id.TryCpid());
+
+            uint256 existing;
+            if (pool.HasMRCForCpid(cpid, &existing)) {
+                return state.DoS(25, error("%s: MRC contract in tx %s has the same CPID as an existing transaction "
+                                           "in the memory pool, %s.",
+                                           __func__,
+                                           tx.GetHash().ToString(),
+                                           existing.ToString()),
+                                 "mrc-duplicate-cpid");
+            }
+
+            tx_contains_valid_mrc = true;
+        }
+    }
+
+    // Check for conflicts with in-memory transactions
+    CTransaction* ptxOld = nullptr;
+    {
+        LOCK(pool.cs); // protect pool.mapNextTx
+        for (unsigned int i = 0; i < tx.vin.size(); i++)
+        {
+            COutPoint outpoint = tx.vin[i].prevout;
+            if (pool.mapNextTx.count(outpoint))
+            {
+                // Disable replacement feature for now
+                return false;
+
+                // Allow replacing with a newer version of the same transaction
+                if (i != 0)
+                    return false;
+                ptxOld = pool.mapNextTx[outpoint].ptx;
+                if (IsFinalTx(*ptxOld))
+                    return false;
+                if (!tx.IsNewerThan(*ptxOld))
+                    return false;
+                for (unsigned int i = 0; i < tx.vin.size(); i++)
+                {
+                    COutPoint outpoint = tx.vin[i].prevout;
+                    if (!pool.mapNextTx.count(outpoint) || pool.mapNextTx[outpoint].ptx != ptxOld)
+                        return false;
+                }
+                break;
+            }
+        }
+    }
+
+    // Computed inside the CTxDB scope below and reused when building the
+    // mempool entry after acceptance.
+    CAmount nFees = 0;
+    unsigned int nSize = 0;
+
+    {
+        CTxDB txdb("r");
+
+        // do we already have it?
+        if (txdb.ContainsTx(hash))
+            return false;
+
+        MapPrevTx mapInputs;
+        std::map<uint256, CTxIndex> mapUnused;
+        bool fInvalid = false;
+        if (!FetchInputs(tx, state, txdb, mapUnused, false, false, mapInputs, fInvalid))
+        {
+            if (fInvalid)
+                return error("AcceptToMemoryPool : FetchInputs found invalid tx %s", hash.ToString().substr(0,10).c_str());
+            if (pfMissingInputs)
+                *pfMissingInputs = true;
+            return false;
+        }
+
+        // Check for non-standard pay-to-script-hash in inputs
+        if (!AreInputsStandard(tx, mapInputs) && !fTestNet)
+            return state.Invalid(error("AcceptToMemoryPool : nonstandard transaction input"),
+                                 "bad-txns-nonstandard-inputs");
+
+        // Note: if you modify this code to accept non-standard transactions, then
+        // you should add code here to check that the transaction does a
+        // reasonable number of ECDSA signature verifications.
+
+        nFees = GetValueIn(tx, mapInputs) - tx.GetValueOut();
+        nSize = ::GetSerializeSize(tx, SER_NETWORK, PROTOCOL_VERSION);
+
+        // Expose the computed fee/size for callers like testmempoolaccept.
+        if (fee_out) *fee_out = nFees;
+        if (vsize_out) *vsize_out = nSize;
+
+        // Don't accept it if it can't get into a block
+        CAmount txMinFee = GetMinFee(tx, 1000, GMF_RELAY, nSize);
+        if (nFees < txMinFee)
+            return state.Invalid(error("AcceptToMemoryPool : not enough fees %s, %" PRId64 " < %" PRId64 ", nSize %" PRId64,
+                                       hash.ToString().c_str(),
+                                       nFees, txMinFee, nSize),
+                                 "insufficient-fee");
+
+        // Validate any contracts published in the transaction:
+        if (!tx.GetContracts().empty() && !CheckContracts(tx, state, mapInputs, pindexBest->nHeight)) {
+            return false;
+        }
+
+        // BIP68: Check relative lock-time (sequence locks) when v14 is active.
+        if (IsV14Enabled(nBestHeight + 1)) {
+            if (!CheckSequenceLocks(tx, 0, pindexBest)) {
+                return state.Invalid(error("AcceptToMemoryPool : sequence locks not satisfied for tx %s",
+                                           hash.ToString().substr(0, 10).c_str()),
+                                     "non-BIP68-final");
+            }
+        }
+
+        // Check against previous transactions
+        // This is done last to help prevent CPU exhaustion denial-of-service attacks.
+        if (!ConnectInputs(tx, state, txdb, mapInputs, mapUnused, CDiskTxPos(1,1,1), pindexBest, false, false))
+        {
+            LogPrint(BCLog::LogFlags::MEMPOOL, "WARNING: %s: Unable to Connect Inputs %s.",
+                     __func__,
+                     hash.ToString().c_str());
+
+            return false;
+        }
+
+        // If we accepted a transaction with a valid mrc contract, then signal MRC changed.
+        if (tx_contains_valid_mrc && !test_only) {
+            uiInterface.MRCChanged();
+        }
+    }
+
+    // testmempoolaccept: all checks have passed; report acceptance without
+    // mutating the pool or notifying wallets.
+    if (test_only) {
+        return true;
+    }
+
+    // Store transaction in memory
+    CTransactionRef replaced_tx; // copy of the replaced tx, captured before remove() frees it
+    {
+        LOCK(pool.cs);
+        if (ptxOld)
+        {
+            LogPrint(BCLog::LogFlags::MEMPOOL, "AcceptToMemoryPool : replacing tx %s with new version", ptxOld->GetHash().ToString());
+            // ptxOld points into the mempool entry that remove() erases below, so
+            // copy the transaction out first; the wallet-erase signal further down
+            // must use this copy, not the dangling ptxOld.
+            replaced_tx = MakeTransactionRef(*ptxOld);
+            pool.remove(*ptxOld);
+        }
+        // entry_time is non-zero only when reloading from unbroadcast.dat: preserve
+        // the original pool-entry time so tx age and eviction ordering survive a
+        // restart. In practice this only affects the node's non-wallet reloaded txs;
+        // a wallet tx is re-pooled with a fresh time by ReacceptWalletTransactions
+        // before LoadUnbroadcast runs, so its persisted time is not applied.
+        CTxMemPoolEntry entry(tx, nFees, entry_time != 0 ? entry_time : GetAdjustedTime(),
+                              nBestHeight, nSize);
+        pool.addUnchecked(hash, entry);
+
+        // Enforce the -maxmempool cap. Protect the transaction we just accepted so
+        // it is not immediately self-evicted: given Gridcoin's flat feerate and the
+        // newest-first tie-break it would otherwise always be the first victim,
+        // failing the node's own sends. Room is made by evicting other, lower-
+        // priority entries instead. If the pool is still over the cap afterwards,
+        // this transaction alone exceeds it -- only possible with an unreasonably
+        // small -maxmempool, which the init-time floor already rejects; treat that
+        // as a "mempool full" rejection.
+        pool.TrimToSize(pool.GetMaxSize(), /*removed=*/nullptr, /*protect=*/&hash);
+        if (!pool.exists(hash)) {
+            return state.Invalid(error("AcceptToMemoryPool : transaction %s rejected, mempool full",
+                                       hash.ToString()),
+                                 "mempool-full");
+        }
+    }
+
+    // Notify the validation-signal layer that the transaction was added to
+    // mempool, so subscribers (e.g. the wallet) can show it immediately with
+    // 0 confirmations. Emitted synchronously under cs_main (held on entry),
+    // preserving the cs_main -> signals -> cs_wallet order.
+    GetMainSignals().TransactionAddedToMempool(MakeTransactionRef(tx));
+
+    ///// are we sure this is ok when loading transactions or restoring block txes
+    // If updated, erase old tx from wallet
+    if (replaced_tx)
+    {
+        // The replaced tx was removed from the mempool above; notify the
+        // validation-signal layer so the wallet drops its now-defunct copy.
+        // Emitted synchronously under cs_main (held on entry), preserving the
+        // cs_main -> signals -> cs_wallet order and replacing the legacy
+        // cs_setpwalletRegistered EraseFromWallets wrapper (issue #3030). Uses the
+        // copy captured before remove() -- ptxOld dangles once the pool entry is gone.
+        GetMainSignals().TransactionReplacedInMempool(replaced_tx);
+    }
+
+    LogPrint(BCLog::LogFlags::MEMPOOL, "AcceptToMemoryPool : accepted %s (poolsz %" PRIszu ")", hash.ToString(), pool.mapTx.size());
+
+    return true;
 }
